auth_server/requirements.txt
numpy
opencv-python
tornado
imutils
scikit-learn
psycopg2-binary
bcrypt